requirements.txt
pytest==7.3.2
selene==2.0.0rc4
allure-pytest==2.13.5
pytest-testdox==3.1.0
pytest-env==1.0.1
